requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.6
psycopg2-binary==2.9.5
sqlparse==0.4.4
typing_extensions==4.6.3
